backend/requirements.txt
boto3==1.13.1
botocore==1.16.1
PyJWT==1.7.1
python-jose==3.1.0
lambda-decorators==0.6.0